+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ in development
 * Fix a bug with a user inside the context of the live action which was created using alias
   execution endpoint incorrectly being set to the system user (``stanley``) instead of the
   authenticated user which triggered the execution. (bug fix)
+* Include state_info for Mistral workflow and tasks in the action execution result. (improvement)
 * Introduce a new ``timeout`` action execution status which represents an action execution
   timeout. Previously, executions which timed out had status set to ``failure``. Keep in mind
   that timeout is just a special type of a failure. (new feature)